perf(sources): tree-byte adaptive budget for the census spill decoded layer
Live whale-page telemetry still showed spill_load ~120s: whale trees blow
past the decoded layer's fixed 256MiB payload-denominated budget and fall
back to pickle decode. Move estimate_parsed_tree_bytes (#3209) to a
layering-neutral home (polylogue/pipeline/parsed_tree_size.py -- sources
cannot import daemon) and budget the decoded layer by ESTIMATED TREE
BYTES with an adaptive RAM/16 budget clamped [256MiB, 2GiB], FIFO
eviction, whales-never-retained. Also raises effective census parse
parallelism at rebuild time via POLYLOGUE_INGEST_PARSE_WORKERS on the
runner (default cap is min(8, cpus-1); this machine has 24 threads).
