+++ docs/topology-status.md
@@ -28,12 +28,12 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 893
+- **Stable** (no move scoped): 894
 - **Kernel** (polylogue/ root): 8
 - **Primitives** (storage-root): 19
 - **TBD** (cell needs explicit assignment): 9
-- **Total declared**: 1064
-- **Realized polylogue/**/*.py**: 1064 files declared
+- **Total declared**: 1065
+- **Realized polylogue/**/*.py**: 1065 files declared
 
 ### TBD cells (require explicit routing)
